Share timezone label formatting
