+++ .changeset/share-timezone-utilities.md
@@ -2,4 +2,4 @@
 '@prairielearn/utils': minor
 ---
 
-Add shared timezone interpretation, local-calendar, and timezone-catalog utilities under `@prairielearn/utils/timezone`.
+Add shared timezone interpretation, formatting, local-calendar, and timezone-catalog utilities under `@prairielearn/utils/timezone`.
